fix(evals): delete orphan metrics_calculator.py, consolidate into eval_runner.MetricsCalculator
metrics_calculator.py duplicated eval_runner.py's MetricsCalculator with no
Python caller anywhere in the tree - only docs pointed at it - and that
duplication was the root cause of the calculate_semantic_similarity drift
fixed in TASK-862. Deletes the orphan, moves its 3 unique methods
(calculate_rouge_scores, calculate_classification_metrics,
calculate_all_metrics) into eval_runner.py's MetricsCalculator, repoints
README.md/DEVELOPER_GUIDE.md/REFACTORING_COMPLETE.md at the survivor, and
adds an AST-based regression guard asserting exactly one MetricsCalculator
class exists in the Evals package.

Closes TASK-863.
